Load_files/requirements.txt
celery==5.3.6
Django==4.2.9
djangorestframework==3.14.0
psycopg==3.1.16
psycopg2-binary==2.9.9
redis==5.0.1
python-decouple==3.8
requests==2.31.0
drf-yasg==1.21.7
django-celery-beat==2.5.0
Pillow